refactor: extract season and album picker mixins (task 3.1)
Move season picker (2 methods) and album picker (3 methods) into
SeasonPickerMixin and AlbumPickerMixin classes. MediaHandler inherits
from both mixins. No test changes needed.

handler.py: 1091 -> 734 lines. All 1636 tests pass.
